fifififinal lastlastl lalalastststst perfectttttoer oasija n;a donedonedodododnenenen

+++ BlackJack/Main.cpp
@@ -39,7 +39,7 @@ int main()
 				cout << "WhattoDo:" << OptiontoString.at(temp.first);
 				cout << "\tExpectedValue:" << temp.second << endl;
 				outputfile << "WhattoDo:" << OptiontoString.at(temp.first) << "\t" << "ExpectedValue:" << temp.second << endl;
-				sum += temp.second * (i == j ? 1 : 2) * (deck.count(i) * deck.count(j) * deck.count(k));
+				sum += temp.second * (i == j ? 1 : 2) * (double)(deck.count(i) * deck.count(j) * deck.count(k));
 			}
 		}
 	}
